refactor/PRI-93 Home crean up
